pyproject.toml
[tool.poetry]
name = "nim-image-to-recipe"
version = "0.1.0"
description = ""
authors = ["ttback"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.2.5"
gradio = "^4.36.1"
langgraph = "^0.0.69"
langchain-nvidia-ai-endpoints = "^0.1.2"
python-dotenv = "^1.0.1"
openai = "^1.34.0"
httpx = "^0.27.0"
transformers = "^4.41.2"
langchain-chroma = "^0.1.1"
langchain-community = "^0.2.5"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
tiktoken = "^0.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
